rename types file
